+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 #### Features Added
 
-* Added a new non-streaming OrderBy query pipeline and a query feature`NonStreamingOrderBy` to support Vector Search queries. - See [PR 39897](https://github.com/Azure/azure-sdk-for-java/pull/39897/) 
+* Added support for non-streaming OrderBy query and a query feature `NonStreamingOrderBy` to support Vector Search queries. - See [PR 39897](https://github.com/Azure/azure-sdk-for-java/pull/39897/) 
 
 #### Breaking Changes